Resolve bestImage conflicts with a counter
- Only the latest rendering/palette change is accepted
- This eliminates the dependency on the fact that `MultithreadedExecutor` works sequentially

However, an incorrect image being shown could still be caused, when the counter overflows: user scrolls up, causing a new re-render; then one pool thread is suspended, and the user scrolls up Integer.MAX_VALUE * 2 times, so that the counter overflows and is now smaller than the one of the suspended thread; the latter is then resumed, causing an old image to be shown. This is not a regression though; previously, only one scroll + suspend would have been needed.
A thread not being 4 billion times slower than the rest is probably a reasonable assumption, though.